infra/serializers/user_serializer.py
from rest_framework import serializers
from app.models import CustomUser


class CustomUserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = ["id", "username", "profile_picture"]
