requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==11.1.0
PyJWT==2.10.0
sqlparse==0.5.0
typing_extensions==4.12.2
